+++ w32023-to-w3up.test.js
@@ -5,7 +5,7 @@ import * as CAR from "@ucanto/transport/car"
 import * as Client from '@ucanto/client'
 import * as ed25519 from '@ucanto/principal/ed25519'
 import * as Server from "@ucanto/server"
-import { migrate } from './migrate-w32023-to-w3up.js'
+import { migrate } from './w32023-to-w3up.js'
 import { IncomingMessage, createServer } from 'http'
 import { MapCidToPromiseResolvers } from './promise.js'
 import { ReadableStream, TransformStream } from 'stream/web'
